+++ packages/bundles/src/react-refresh-runtime.ts
@@ -1,0 +1,3 @@
+import runtime from 'react-refresh/runtime';
+
+export default runtime;